backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
gunicorn==21.2.0